app/core/security.py
import jwt
from datetime import datetime, timedelta
from app.core.config import settings
from sqlalchemy.orm import Session
from app.db.models import RefreshToken

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)

def create_refresh_token(data: dict, db: Session):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(days=settings.REFRESH_TOKEN_EXPIRE_DAYS)
    to_encode.update({"exp": expire})
    refresh_token = jwt.encode(to_encode, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)

    token = RefreshToken(
        user_id=data["user_id"],
        refresh_token=refresh_token,
        expires_at=expire,
        is_revoked=False
    )
    db.add(token)
    db.commit()
    return refresh_token

def decode_token(token: str):
    return jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
