requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
bcrypt
pyjwt
pydantic
requests
python-dotenv
